[fixed] Loot group chances are now doublespinbox. [fixed] Entries in model viewer dropdown alphabetized. [added] some icons to the resources
git-svn-id: http://svn2.assembla.com/svn/swgemu/tools@3911 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ WorldSpawnerTool/lootmanager.h
@@ -20,7 +20,7 @@ public:
     ~LootManager();
 
     void insertTemplateToTable(QExplicitlySharedDataPointer<LootItemTemplate>);
-    void insertEntryToGroupTable(const QString&, int);
+    void insertEntryToGroupTable(const QString&, double);
     int insertGroupToComboBox(QExplicitlySharedDataPointer<LootGroup>);
 
 public slots:
